Changed it to more class structured
