+++ src/main/java/LevelController.java
@@ -1,14 +1,11 @@
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.canvas.Canvas;
-import javafx.scene.canvas.GraphicsContext;
-import javafx.scene.image.Image;
 import javafx.scene.layout.StackPane;
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
+import java.io.File;
 import java.net.URL;
+import java.util.ArrayList;
 import java.util.ResourceBundle;
 
 /**
@@ -21,60 +18,49 @@ public class LevelController implements Initializable {
     @FXML
     Canvas canvas;
 
-    protected static final int num_rows = 26;
-    protected static final int num_cols = 26;
-    protected static Integer map[][];
+    private ArrayList<String> maps;
+    private int currLvl;
 
     @Override
     public void initialize(URL location, ResourceBundle resources) {
-        GraphicsContext gc = canvas.getGraphicsContext2D();
-        drawMap(gc);
+
+        //Scan folder for maps
+        //load first map
+        //otherwise output no maps found
+
+        maps = new ArrayList<>();
+        findMaps();
+
+        if(maps.size() > 0) {
+            //load first map
+            currLvl = 0;
+            createLvl();
+        }else{
+            System.out.println("No maps found!");
+        }
+
     }
 
-    public void drawMap(GraphicsContext gc){
-        Image image = new Image(getClass().getResourceAsStream("BubbleBobbleWall32b.png"));
-        readMap();
-        System.out.println(map[0][0]);
-        for(int row = 0; row < num_rows; row++) {
-            for(int col = 0; col < num_cols; col++) {
-                if(map[row][col] == 1){
-                    gc.drawImage(image, col*32, row*32);
+    private void findMaps() {
+        File folder = new File("c:/users/jim/workspace/bubblebobble/src/main/resources");
+        File[] listOfFiles = folder.listFiles();
+        assert listOfFiles != null;
+        for (File listOfFile : listOfFiles) {
+            if (listOfFile.isFile()) {
+                if (listOfFile.getName().matches("map[0-9]*.txt")) {
+                    maps.add(listOfFile.getName());
                 }
             }
         }
     }
 
-    public void readMap(){
-        int row = 0;
-        map = new Integer[num_rows][num_cols];
-
-        BufferedReader reader = null;
-
-        try {
-            reader = new BufferedReader(new InputStreamReader(getClass().getResourceAsStream("map1.txt")));
-            String line;
-            while ((line = reader.readLine()) != null) {
-                String[] cols = line.split(" ");
-                if(cols.length == num_cols) {
-                    for(int colum = 0; colum < cols.length; colum++) {
-                        map[row][colum] = Integer.parseInt(cols[colum]);
-                    }
-                }
-                row++;
-            }
-        }
-        catch (IOException e) {
-            e.printStackTrace();
-        }
-        finally {
-            if(reader != null) {
-                try {
-                    reader.close();
-                }
-                catch (IOException e) {
-                    e.printStackTrace();
-                }
-            }
-        }
+    public void createLvl() {
+        new Level(maps.get(currLvl), canvas);
     }
+
+    public void nextLevel() {
+        currLvl++;
+        createLvl();
+    }
+
 }
